requirements.txt
# Main training requirements
tensorflow_gpu == 1.13.1
numpy == 1.15.4
progressbar2
pandas
six
pyxdg
attrdict

# Requirements for building native_client files
setuptools

# Requirements for importers
sox
bs4
requests
librosa
soundfile

# Miscellaneous scripts
paramiko >= 2.1
scipy
matplotlib
